Added Independent File for BitmaskUnpack
